requirements.txt
Flask==3.0.3
Werkzeug>=3.0.0
pymysql==1.1.0
pdfminer.six==20221105
pyresparser==1.0.6
geopy==2.4.1
geocoder==1.38.1
pandas==2.2.2
plotly==5.22.0
Pillow==10.4.0
nltk==3.8.1
pdfminer3
PyPDF2
PyMuPDF
sentence-transformers